backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
aiofiles
argon2-cffi
cloudinary
imagehash
Pillow
cryptography
numpy
scikit-learn
sentence-transformers
fpdf2
